refactor: Refactor UMI package to use Solana-specific logic and data types
- Renamed the `UmiMessage.kt` file to `SolanaMessage.kt` and made necessary updates to package declaration, class name, and serialization logic.
- Updated project names and included additional projects in `settings.gradle.kts`.
- Renamed the `build.gradle.kts` file to `solana/build.gradle.kts`.
- Renamed the `MessageTest.kt` file to `solana/src/commonTest/kotlin/com/metaplex/solana/MessageTest.kt` and made necessary updates to package import, function names, and usage of `SolanaTransactionBuilder`.
- Renamed the `UmiTransactionBuilder.kt` file to `SolanaTransactionBuilder.kt` and made necessary updates to package declaration, class name, transaction type, and comments.
- Renamed the `Shortvec.kt` file to `solana/src/commonMain/kotlin/com/metaplex/solana/Shortvec.kt` and updated the package declaration.
- Renamed the `UmiTransaction.kt` file to `SolanaTransaction.kt`, updated package declaration, class name, references to `UmiMessage`, and function names. Fixed a typo in a comment.
